vitalio/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-autoslug==1.9.8
django-crispy-forms==1.11.0
django-discover-runner==1.0
django-environ==0.4.5
django-jenkins==0.110.0
django-role-permissions==3.1.1
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.52.0
urllib3==1.26.3
